refactor: move config code
