+++ embrace-android-config/src/main/kotlin/io/embrace/android/embracesdk/internal/config/source/CombinedRemoteConfigSource.kt
@@ -2,7 +2,6 @@ package io.embrace.android.embracesdk.internal.config.source
 
 import io.embrace.android.embracesdk.internal.config.remote.RemoteConfig
 import io.embrace.android.embracesdk.internal.config.store.RemoteConfigStore
-import io.embrace.android.embracesdk.internal.utils.EmbTrace
 import io.embrace.android.embracesdk.internal.worker.BackgroundWorker
 import java.util.concurrent.TimeUnit
 
@@ -17,22 +16,18 @@ class CombinedRemoteConfigSource(
 
     // the remote config that is used for the lifetime of the process.
     private val response by lazy {
-        EmbTrace.trace("load-config-from-store") {
-            store.loadResponse()
-        }
+        store.loadResponse()
     }
 
     fun getConfig(): RemoteConfig? = response?.cfg
 
     fun scheduleConfigRequests() {
-        EmbTrace.trace("schedule-http-request") {
-            worker.scheduleWithFixedDelay(
-                ::attemptConfigRequest,
-                0,
-                intervalMs,
-                TimeUnit.MILLISECONDS
-            )
-        }
+        worker.scheduleWithFixedDelay(
+            ::attemptConfigRequest,
+            0,
+            intervalMs,
+            TimeUnit.MILLISECONDS
+        )
     }
 
     private fun attemptConfigRequest() {
