+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/config/ConfigModuleImpl.kt
@@ -1,9 +1,7 @@
-package io.embrace.android.embracesdk.internal.injection
+package io.embrace.android.embracesdk.internal.config
 
 import io.embrace.android.embracesdk.internal.comms.api.ApiUrlBuilder
 import io.embrace.android.embracesdk.internal.comms.api.EmbraceApiUrlBuilder
-import io.embrace.android.embracesdk.internal.config.ConfigService
-import io.embrace.android.embracesdk.internal.config.ConfigServiceImpl
 import io.embrace.android.embracesdk.internal.config.source.CombinedRemoteConfigSource
 import io.embrace.android.embracesdk.internal.config.source.OkHttpRemoteConfigSource
 import io.embrace.android.embracesdk.internal.config.source.RemoteConfigSource
@@ -11,14 +9,16 @@ import io.embrace.android.embracesdk.internal.config.store.RemoteConfigStore
 import io.embrace.android.embracesdk.internal.config.store.RemoteConfigStoreImpl
 import io.embrace.android.embracesdk.internal.envelope.BuildInfo
 import io.embrace.android.embracesdk.internal.envelope.CpuAbi
+import io.embrace.android.embracesdk.internal.injection.CoreModule
+import io.embrace.android.embracesdk.internal.injection.InitModule
+import io.embrace.android.embracesdk.internal.injection.OpenTelemetryModule
+import io.embrace.android.embracesdk.internal.injection.WorkerThreadModule
+import io.embrace.android.embracesdk.internal.injection.singleton
 import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.symbols.SymbolServiceImpl
 import io.embrace.android.embracesdk.internal.utils.EmbTrace
 import io.embrace.android.embracesdk.internal.utils.Uuid.getEmbUuid
 import io.embrace.android.embracesdk.internal.worker.Worker
-import okhttp3.OkHttpClient
-import okhttp3.Protocol
 import java.io.File
-import java.util.concurrent.TimeUnit
 
 class ConfigModuleImpl(
     initModule: InitModule,
@@ -28,20 +28,7 @@ class ConfigModuleImpl(
 ) : ConfigModule {
 
     companion object {
-        private const val DEFAULT_CONNECTION_TIMEOUT_SECONDS = 10L
-        private const val DEFAULT_READ_TIMEOUT_SECONDS = 60L
         private const val DEVICE_IDENTIFIER_KEY = "io.embrace.deviceid"
-    }
-
-    override val okHttpClient by singleton {
-        EmbTrace.trace("okhttp-client-init") {
-            OkHttpClient()
-                .newBuilder()
-                .protocols(listOf(Protocol.HTTP_2, Protocol.HTTP_1_1))
-                .connectTimeout(DEFAULT_CONNECTION_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-                .readTimeout(DEFAULT_READ_TIMEOUT_SECONDS, TimeUnit.SECONDS)
-                .build()
-        }
     }
 
     override val combinedRemoteConfigSource: CombinedRemoteConfigSource? by singleton {
@@ -67,7 +54,7 @@ class ConfigModuleImpl(
     private val remoteConfigSource: RemoteConfigSource? by singleton {
         if (initModule.onlyOtelExportEnabled()) return@singleton null
         OkHttpRemoteConfigSource(
-            okhttpClient = okHttpClient,
+            okhttpClient = initModule.okHttpClient,
             apiUrlBuilder = urlBuilder ?: return@singleton null,
             serializer = initModule.jsonSerializer,
         )
