+++ embrace-android-core/src/test/kotlin/io/embrace/android/embracesdk/internal/injection/ConfigModuleImplTest.kt
@@ -5,6 +5,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.embrace.android.embracesdk.fakes.FakeOpenTelemetryModule
 import io.embrace.android.embracesdk.fakes.injection.FakeInitModule
 import io.embrace.android.embracesdk.fakes.injection.FakeWorkerThreadModule
+import io.embrace.android.embracesdk.internal.config.ConfigModuleImpl
 import org.junit.Assert.assertNotNull
 import org.junit.Test
 import org.junit.runner.RunWith
